+++ docs/demos/data.ts
@@ -1,788 +1,4 @@
 const datas = {
-  'line/base': [
-    {
-      date: '2017-06-05',
-      value: 116,
-    },
-    {
-      date: '2017-06-06',
-      value: 129,
-    },
-    {
-      date: '2017-06-07',
-      value: 135,
-    },
-    {
-      date: '2017-06-08',
-      value: 86,
-    },
-    {
-      date: '2017-06-09',
-      value: 73,
-    },
-    {
-      date: '2017-06-10',
-      value: 85,
-    },
-    {
-      date: '2017-06-11',
-      value: 73,
-    },
-    {
-      date: '2017-06-12',
-      value: 68,
-    },
-    {
-      date: '2017-06-13',
-      value: 92,
-    },
-    {
-      date: '2017-06-14',
-      value: 130,
-    },
-    {
-      date: '2017-06-15',
-      value: 245,
-    },
-    {
-      date: '2017-06-16',
-      value: 139,
-    },
-    {
-      date: '2017-06-17',
-      value: 115,
-    },
-    {
-      date: '2017-06-18',
-      value: 111,
-    },
-    {
-      date: '2017-06-19',
-      value: 309,
-    },
-    {
-      date: '2017-06-20',
-      value: 206,
-    },
-    {
-      date: '2017-06-21',
-      value: 137,
-    },
-    {
-      date: '2017-06-22',
-      value: 128,
-    },
-    {
-      date: '2017-06-23',
-      value: 85,
-    },
-    {
-      date: '2017-06-24',
-      value: 94,
-    },
-    {
-      date: '2017-06-25',
-      value: 71,
-    },
-    {
-      date: '2017-06-26',
-      value: 106,
-    },
-    {
-      date: '2017-06-27',
-      value: 84,
-    },
-    {
-      date: '2017-06-28',
-      value: 93,
-    },
-    {
-      date: '2017-06-29',
-      value: 85,
-    },
-    {
-      date: '2017-06-30',
-      value: 73,
-    },
-    {
-      date: '2017-07-01',
-      value: 83,
-    },
-    {
-      date: '2017-07-02',
-      value: 125,
-    },
-    {
-      date: '2017-07-03',
-      value: 107,
-    },
-    {
-      date: '2017-07-04',
-      value: 82,
-    },
-    {
-      date: '2017-07-05',
-      value: 44,
-    },
-    {
-      date: '2017-07-06',
-      value: 72,
-    },
-    {
-      date: '2017-07-07',
-      value: 106,
-    },
-    {
-      date: '2017-07-08',
-      value: 107,
-    },
-    {
-      date: '2017-07-09',
-      value: 66,
-    },
-    {
-      date: '2017-07-10',
-      value: 91,
-    },
-    {
-      date: '2017-07-11',
-      value: 92,
-    },
-    {
-      date: '2017-07-12',
-      value: 113,
-    },
-    {
-      date: '2017-07-13',
-      value: 107,
-    },
-    {
-      date: '2017-07-14',
-      value: 131,
-    },
-    {
-      date: '2017-07-15',
-      value: 111,
-    },
-    {
-      date: '2017-07-16',
-      value: 64,
-    },
-    {
-      date: '2017-07-17',
-      value: 69,
-    },
-    {
-      date: '2017-07-18',
-      value: 88,
-    },
-    {
-      date: '2017-07-19',
-      value: 77,
-    },
-    {
-      date: '2017-07-20',
-      value: 83,
-    },
-    {
-      date: '2017-07-21',
-      value: 111,
-    },
-    {
-      date: '2017-07-22',
-      value: 57,
-    },
-    {
-      date: '2017-07-23',
-      value: 55,
-    },
-    {
-      date: '2017-07-24',
-      value: 60,
-    },
-  ],
-  'line/withPoint': [
-    {
-      day: '周一',
-      value: 300,
-    },
-    {
-      day: '周二',
-      value: 400,
-    },
-    {
-      day: '周三',
-      value: 350,
-    },
-    {
-      day: '周四',
-      value: 500,
-    },
-    {
-      day: '周五',
-      value: 490,
-    },
-    {
-      day: '周六',
-      value: 600,
-    },
-    {
-      day: '周日',
-      value: 900,
-    },
-  ],
-  'line/smooth': [
-    {
-      time: '2016-08-08 00:00:00',
-      tem: 10,
-    },
-    {
-      time: '2016-08-08 00:10:00',
-      tem: 22,
-    },
-    {
-      time: '2016-08-08 00:30:00',
-      tem: 20,
-    },
-    {
-      time: '2016-08-09 00:35:00',
-      tem: 26,
-    },
-    {
-      time: '2016-08-09 01:00:00',
-      tem: 20,
-    },
-    {
-      time: '2016-08-09 01:20:00',
-      tem: 26,
-    },
-    {
-      time: '2016-08-10 01:40:00',
-      tem: 28,
-    },
-    {
-      time: '2016-08-10 02:00:00',
-      tem: 20,
-    },
-    {
-      time: '2016-08-10 02:20:00',
-      tem: 18,
-    },
-  ],
-  'line/pan': [
-    {
-      title: 'Bohemian Rhapsody',
-      artist: 'Queen',
-      release: 1975,
-      year: '1999',
-      rank: '1',
-      count: 978,
-    },
-    {
-      title: 'Hotel California',
-      artist: 'Eagles',
-      release: 1977,
-      year: '1999',
-      rank: '2',
-      count: 1284,
-    },
-    {
-      title: 'Child In Time',
-      artist: 'Deep Purple',
-      release: 1972,
-      year: '1999',
-      rank: '3',
-      count: 1117,
-    },
-    {
-      title: 'Stairway To Heaven',
-      artist: 'Led Zeppelin',
-      release: 1971,
-      year: '1999',
-      rank: '4',
-      count: 1132,
-    },
-    {
-      title: 'Paradise By The Dashboard Light',
-      artist: 'Meat Loaf',
-      release: 1978,
-      year: '1999',
-      rank: '5',
-      count: 1187,
-    },
-    {
-      title: 'Yesterday',
-      artist: 'The Beatles',
-      release: 1965,
-      year: '1999',
-      rank: '6',
-      count: 909,
-    },
-    {
-      title: 'Angie',
-      artist: 'The Rolling Stones',
-      release: 1973,
-      year: '1999',
-      rank: '8',
-      count: 1183,
-    },
-    {
-      title: 'Bridge Over Troubled Water',
-      artist: 'Simon & Garfunkel',
-      release: 1970,
-      year: '1999',
-      rank: '9',
-      count: 1111,
-    },
-    {
-      title: 'A Whiter Shade Of Pale',
-      artist: 'Procol Harum',
-      release: 1967,
-      year: '1999',
-      rank: '10',
-      count: 1190,
-    },
-    {
-      title: 'Hey Jude',
-      artist: 'The Beatles',
-      release: 1968,
-      year: '1999',
-      rank: '11',
-      count: 1037,
-    },
-    {
-      title: 'House Of The Rising Sun',
-      artist: 'The Animals',
-      release: 1964,
-      year: '1999',
-      rank: '13',
-      count: 543,
-    },
-    {
-      title: 'Goodnight Saigon',
-      artist: 'Billy Joel',
-      release: 1983,
-      year: '1999',
-      rank: '14',
-      count: 748,
-    },
-    {
-      title: 'Dancing Queen',
-      artist: 'ABBA',
-      release: 1976,
-      year: '1999',
-      rank: '16',
-      count: 1111,
-    },
-    {
-      title: 'Another Brick In The Wall',
-      artist: 'Pink Floyd',
-      release: 1979,
-      year: '1999',
-      rank: '17',
-      count: 1266,
-    },
-    {
-      title: 'Sunday Bloody Sunday',
-      artist: 'U2',
-      release: 1985,
-      year: '1999',
-      rank: '18',
-      count: 1087,
-    },
-    {
-      title: 'Tears In Heaven',
-      artist: 'Eric Clapton',
-      release: 1992,
-      year: '1999',
-      rank: '21',
-      count: 435,
-    },
-    {
-      title: 'Old And Wise',
-      artist: 'The Alan Parsons Project',
-      release: 1982,
-      year: '1999',
-      rank: '24',
-      count: 945,
-    },
-    {
-      title: 'Losing My Religion',
-      artist: 'R.E.M.',
-      release: 1991,
-      year: '1999',
-      rank: '25',
-      count: 415,
-    },
-    {
-      title: 'School',
-      artist: 'Supertramp',
-      release: 1974,
-      year: '1999',
-      rank: '26',
-      count: 1011,
-    },
-    {
-      title: 'Who Wants To Live Forever',
-      artist: 'Queen',
-      release: 1986,
-      year: '1999',
-      rank: '30',
-      count: 836,
-    },
-    {
-      title: 'Everybody Hurts',
-      artist: 'R.E.M.',
-      release: 1993,
-      year: '1999',
-      rank: '31',
-      count: 301,
-    },
-    {
-      title: 'Over De Muur',
-      artist: 'Klein Orkest',
-      release: 1984,
-      year: '1999',
-      rank: '32',
-      count: 1166,
-    },
-    {
-      title: 'Paint It Black',
-      artist: 'The Rolling Stones',
-      release: 1966,
-      year: '1999',
-      rank: '33',
-      count: 1077,
-    },
-    {
-      title: 'The Winner Takes It All',
-      artist: 'ABBA',
-      release: 1980,
-      year: '1999',
-      rank: '35',
-      count: 926,
-    },
-    {
-      title: 'Candle In The Wind (1997)',
-      artist: 'Elton John',
-      release: 1997,
-      year: '1999',
-      rank: '37',
-      count: 451,
-    },
-    {
-      title: 'My Heart Will Go On',
-      artist: 'Celine Dion',
-      release: 1998,
-      year: '1999',
-      rank: '41',
-      count: 415,
-    },
-    {
-      title: 'The River',
-      artist: 'Bruce Springsteen',
-      release: 1981,
-      year: '1999',
-      rank: '48',
-      count: 723,
-    },
-    {
-      title: 'With Or Without You',
-      artist: 'U2',
-      release: 1987,
-      year: '1999',
-      rank: '51',
-      count: 816,
-    },
-    {
-      title: 'Space Oddity',
-      artist: 'David Bowie',
-      release: 1969,
-      year: '1999',
-      rank: '59',
-      count: 1344,
-    },
-    {
-      title: 'Stil In Mij',
-      artist: 'Van Dik Hout',
-      release: 1994,
-      year: '1999',
-      rank: '65',
-      count: 373,
-    },
-    {
-      title: 'Nothing Compares 2 U',
-      artist: "Sinead O'Connor",
-      release: 1990,
-      year: '1999',
-      rank: '90',
-      count: 426,
-    },
-    {
-      title: 'Wonderful Tonight',
-      artist: 'Eric Clapton',
-      release: 1988,
-      year: '1999',
-      rank: '91',
-      count: 515,
-    },
-    {
-      title: 'Blowing In The Wind',
-      artist: 'Bob Dylan',
-      release: 1963,
-      year: '1999',
-      rank: '94',
-      count: 323,
-    },
-    {
-      title: 'Eternal Flame',
-      artist: 'Bangles',
-      release: 1989,
-      year: '1999',
-      rank: '96',
-      count: 495,
-    },
-    {
-      title: 'Non Je Ne Regrette Rien',
-      artist: 'Edith Piaf',
-      release: 1961,
-      year: '1999',
-      rank: '106',
-      count: 178,
-    },
-    {
-      title: 'Con Te Partiro',
-      artist: 'Andrea Bocelli',
-      release: 1996,
-      year: '1999',
-      rank: '109',
-      count: 362,
-    },
-    {
-      title: 'Conquest Of Paradise',
-      artist: 'Vangelis',
-      release: 1995,
-      year: '1999',
-      rank: '157',
-      count: 315,
-    },
-    {
-      title: 'White Christmas',
-      artist: 'Bing Crosby',
-      release: 1954,
-      year: '1999',
-      rank: '218',
-      count: 10,
-    },
-    {
-      title: "(We're gonna) Rock Around The Clock",
-      artist: 'Bill Haley & The Comets',
-      release: 1955,
-      year: '1999',
-      rank: '239',
-      count: 19,
-    },
-    {
-      title: 'Jailhouse Rock',
-      artist: 'Elvis Presley',
-      release: 1957,
-      year: '1999',
-      rank: '247',
-      count: 188,
-    },
-    {
-      title: 'Take Five',
-      artist: 'Dave Brubeck',
-      release: 1962,
-      year: '1999',
-      rank: '279',
-      count: 204,
-    },
-    {
-      title: "It's Now Or Never",
-      artist: 'Elvis Presley',
-      release: 1960,
-      year: '1999',
-      rank: '285',
-      count: 221,
-    },
-    {
-      title: 'Heartbreak Hotel',
-      artist: 'Elvis Presley',
-      release: 1956,
-      year: '1999',
-      rank: '558',
-      count: 109,
-    },
-    {
-      title: 'One Night',
-      artist: 'Elvis Presley',
-      release: 1959,
-      year: '1999',
-      rank: '622',
-      count: 71,
-    },
-    {
-      title: 'Johnny B. Goode',
-      artist: 'Chuck Berry',
-      release: 1958,
-      year: '1999',
-      rank: '714',
-      count: 89,
-    },
-    {
-      title: 'Unforgettable',
-      artist: "Nat 'King' Cole",
-      release: 1951,
-      year: '1999',
-      rank: '1188',
-      count: 20,
-    },
-    {
-      title: 'La Mer',
-      artist: 'Charles Trenet',
-      release: 1952,
-      year: '1999',
-      rank: '1249',
-      count: 24,
-    },
-    {
-      title: 'The Road Ahead',
-      artist: 'City To City',
-      release: 1999,
-      year: '1999',
-      rank: '1999',
-      count: 262,
-    },
-    {
-      title: 'What It Is',
-      artist: 'Mark Knopfler',
-      release: 2000,
-      year: '2000',
-      rank: '545',
-      count: 291,
-    },
-    {
-      title: 'Overcome',
-      artist: 'Live',
-      release: 2001,
-      year: '2001',
-      rank: '879',
-      count: 111,
-    },
-    {
-      title: 'Mooie Dag',
-      artist: 'Blof',
-      release: 2002,
-      year: '2003',
-      rank: '147',
-      count: 256,
-    },
-    {
-      title: 'Clocks',
-      artist: 'Coldplay',
-      release: 2003,
-      year: '2003',
-      rank: '733',
-      count: 169,
-    },
-    {
-      title: 'Sunrise',
-      artist: 'Norah Jones',
-      release: 2004,
-      year: '2004',
-      rank: '405',
-      count: 256,
-    },
-    {
-      title: 'Nine Million Bicycles',
-      artist: 'Katie Melua',
-      release: 2005,
-      year: '2005',
-      rank: '23',
-      count: 250,
-    },
-    {
-      title: 'Rood',
-      artist: 'Marco Borsato',
-      release: 2006,
-      year: '2006',
-      rank: '17',
-      count: 159,
-    },
-    {
-      title: 'If You Were A Sailboat',
-      artist: 'Katie Melua',
-      release: 2007,
-      year: '2007',
-      rank: '101',
-      count: 256,
-    },
-    {
-      title: 'Viva La Vida',
-      artist: 'Coldplay',
-      release: 2009,
-      year: '2009',
-      rank: '11',
-      count: 228,
-    },
-    {
-      title: 'Dochters',
-      artist: 'Marco Borsato',
-      release: 2008,
-      year: '2009',
-      rank: '25',
-      count: 268,
-    },
-    {
-      title: 'Need You Now',
-      artist: 'Lady Antebellum',
-      release: 2010,
-      year: '2010',
-      rank: '210',
-      count: 121,
-    },
-    {
-      title: 'Someone Like You',
-      artist: 'Adele',
-      release: 2011,
-      year: '2011',
-      rank: '6',
-      count: 187,
-    },
-    {
-      title: 'I Follow Rivers',
-      artist: 'Triggerfinger',
-      release: 2012,
-      year: '2012',
-      rank: '79',
-      count: 167,
-    },
-    {
-      title: 'Get Lucky',
-      artist: 'Daft Punk',
-      release: 2013,
-      year: '2013',
-      rank: '357',
-      count: 141,
-    },
-    {
-      title: 'Home',
-      artist: 'Dotan',
-      release: 2014,
-      year: '2014',
-      rank: '82',
-      count: 76,
-    },
-    {
-      title: 'Hello',
-      artist: 'Adele',
-      release: 2015,
-      year: '2015',
-      rank: '23',
-      count: 29,
-    },
-  ],
   'multipleLine/series': [
     {
       date: '2010-01-10',
