+++ docs/demos/line/withPoint/data.ts
@@ -1,0 +1,30 @@
+export default [
+  {
+    day: '周一',
+    value: 300,
+  },
+  {
+    day: '周二',
+    value: 400,
+  },
+  {
+    day: '周三',
+    value: 350,
+  },
+  {
+    day: '周四',
+    value: 500,
+  },
+  {
+    day: '周五',
+    value: 490,
+  },
+  {
+    day: '周六',
+    value: 600,
+  },
+  {
+    day: '周日',
+    value: 900,
+  },
+];
